增删改评论

+++ handler/delete_comment.go
@@ -8,33 +8,35 @@ import (
 	"github.com/labstack/echo/v4"
 )
 
-type deleteCommentReq struct {
-}
-
 type deleteCommentRes struct {
-	Success bool `json:"success"`
+	Code int    `json:"code"`
+	Msg  string `json:"msg"`
 }
 
 // 定义一个接收参数，并返回结果
 func HandleDeleteComment(c echo.Context) error {
-	req := &deleteCommentReq{}
-
-	// 绑定请求参数
-	c.Bind(req)
 	id, _ := strconv.ParseInt(c.Param("id"), 10, 64)
+
+	res := &deleteCommentRes{}
+
+	if id < 0 {
+		res.Code = 10001
+		res.Msg = "无此id"
+		return c.JSON(200, res)
+	}
 
 	// 删除数据库里的数据
 	r, err := dbutil.DB.Exec("delete from comments where id = ?", id)
 	if err != nil {
-		log.Print(err)
-		return err
+		res.Code = 100010
+		res.Msg = err.Error()
+		return c.JSON(200, res)
 	} else {
 		log.Print(r)
 	}
 
-	res := &deleteCommentRes{
-		Success: true,
-	}
+	res.Code = 0
+	res.Msg = "ok"
 	//正确返回相应结果
 	return c.JSON(200, res)
 }
